automated_job_search/requirements.txt
requests
pandas

matplotlib
seaborn

scikit-learn
nltk

python-dotenv
streamlit